% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agriculture.R
\docType{data}
\name{Agriculture}
\alias{Agriculture}
\title{Agricultural financial activity data from Pakistan Social and Living Standards Measurement 2015-16}
\format{A \code{data.table} and  \code{data.frame} with 74718 obs. of  10 variables.
\describe{
       \item{\code{hhcode}}{10 digits Household code.}
       \item{\code{Province}}{Province of Pakistan}
       \item{\code{Region}}{Urban or Rural area of Pakitan}
       \item{\code{Psu}}{08 digits code of primary sampling unit}
       \item{\code{sec}}{Questionair section}
       \item{\code{Code}}{Agicultural activities}
       \item{\code{S7ac1}}{Direct response from respondent}
       \item{\code{S7ac2}}{Direct response from respondent}
       \item{\code{S7ac3}}{Direct response from respondent}
       \item{\code{S7ac4}}{Direct response from respondent}
       \item{\code{S7ac5}}{Direct response from respondent}
       }}
\usage{
data(Agriculture)
}
\description{
\code{Agriculture} Agricultural financial activity data from Pakistan Social and Living Standards Measurement 2015-16.
}
\references{
\enumerate{
\item Pakistan Bureau of Statistics, Micro data (\url{http://www.pbs.gov.pk/}).
 }
}
\seealso{
\code{\link{AssetLiability}}
 , \code{\link{Education}}
 , \code{\link{Employment}}
 , \code{\link{Expenditure}}
 , \code{\link{ICT}}
 , \code{\link{Income}}
 , \code{\link{Land}}
 , \code{\link{LiveStock}}
 , \code{\link{NonAgriculture}}
 , \code{\link{OtherIncome}}
 , \code{\link{PaidReceived}}
 , \code{\link{SurveyInfo}}
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})

Muhammad Arfan Dilber (\email{pbsfsd041@gmail.com})
}
\keyword{datasets}
