% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Land.R
\docType{data}
\name{Land}
\alias{Land}
\title{Owned buildings and land data from Pakistan Social and Living Standards Measurement 2015-16}
\format{A \code{data.table} and  \code{data.frame} with 116421 obs. of  16 variables.
\describe{
       \item{\code{hhcode}}{10 digits Household code.}
       \item{\code{Province}}{Province of Pakistan}
       \item{\code{Region}}{Urban or Rural area of Pakitan}
       \item{\code{Psu}}{08 digits code of primary sampling unit}
       \item{\code{sec}}{Questionair section}
       \item{\code{Code}}{Property type which is owned by household member}
       \item{\code{S6aq01}}{Did any of the HH members own or had owned during the last 1 year any of the following property?}
       \item{\code{S6aq02}}{Is all or part ot this property owned now?}
       \item{\code{S6aq2a}}{if owned how many acres?}
       \item{\code{S6aq3a}}{What was the value of property which was sold during last 1 year?}
       \item{\code{S6aq3b}}{What was the value of property which was Recieved during last 1 year?}
       \item{\code{S6aq3c}}{What was the value of property which was purchased during last 1 year?}
       \item{\code{S6aq3d}}{What was the value of property which was given during last 1 year?}
       \item{\code{S6aq04}}{If rented out, what  was the total net rent received, in cash / kind, during the last 1 year?}
       \item{\code{S6aq05}}{ What was the value of major improvements, renovation and new construction made during the last 1 year?}
       \item{\code{S6aq06}}{ If you wanted to sell the property  how much do you expect to receive(RS)?}
       }}
\usage{
data(Land)
}
\description{
\code{Land} Owned buildings and land data from Pakistan Social and Living Standards Measurement 2015-16.
}
\references{
\enumerate{
\item Pakistan Bureau of Statistics, Micro data (\url{http://www.pbs.gov.pk/}).
 }
}
\seealso{
\code{\link{Agriculture}}
 , \code{\link{AssetLiability}}
 , \code{\link{Education}}
 , \code{\link{Employment}}
 , \code{\link{Expenditure}}
 , \code{\link{ICT}}
 , \code{\link{Income}}
 , \code{\link{LiveStock}}
 , \code{\link{NonAgriculture}}
 , \code{\link{OtherIncome}}
 , \code{\link{PaidReceived}}
 , \code{\link{SurveyInfo}}
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})

Muhammad Arfan Dilber (\email{pbsfsd041@gmail.com})
}
\keyword{datasets}
