% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherIncome.R
\docType{data}
\name{OtherIncome}
\alias{OtherIncome}
\title{Household's other non-permanent income data from Pakistan Social and Living Standards Measurement 2015-16}
\format{A \code{data.table} and  \code{data.frame} with 17379 obs. of  8 variables.
\describe{
       \item{\code{hhcode}}{10 digits Household code.}
       \item{\code{Province}}{Province of Pakistan}
       \item{\code{Region}}{Urban or Rural area of Pakitan}
       \item{\code{Psu}}{08 digits code of primary sampling unit}
       \item{\code{sec}}{Questionair section}
       \item{\code{filler}}{}
       \item{\code{codes}}{Different sources of other income}
       \item{\code{values}}{Amount received from other non-permanent sources of income}
       }}
\usage{
data(OtherIncome)
}
\description{
\code{OtherIncome} Household's other non-permanent income data from Pakistan Social and Living Standards Measurement 2015-16.
}
\references{
\enumerate{
\item Pakistan Bureau of Statistics, Micro data (\url{http://www.pbs.gov.pk/}).
 }
}
\seealso{
\code{\link{Agriculture}}
 , \code{\link{AssetLiability}}
 , \code{\link{Education}}
 , \code{\link{Employment}}
 , \code{\link{Expenditure}}
 , \code{\link{ICT}}
 , \code{\link{Income}}
 , \code{\link{Land}}
 , \code{\link{LiveStock}}
 , \code{\link{NonAgriculture}}
 , \code{\link{PaidReceived}}
 , \code{\link{SurveyInfo}}
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})

Muhammad Arfan Dilber (\email{pbsfsd041@gmail.com})
}
\keyword{datasets}
