% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaidReceived.R
\docType{data}
\name{PaidReceived}
\alias{PaidReceived}
\title{Transfers, received and paid out data from Pakistan Social and Living Standards Measurement 2015-16}
\format{A \code{data.table} and  \code{data.frame} with 168442 obs. of  7 variables.
\describe{
       \item{\code{hhcode}}{10 digits Household code.}
       \item{\code{Province}}{Province of Pakistan}
       \item{\code{Region}}{Urban or Rural area of Pakitan}
       \item{\code{Psu}}{08 digits code of primary sampling unit}
       \item{\code{sec}}{Questionair section}
       \item{\code{itc}}{Sources from which income received or paid out}
       \item{\code{value}}{Received and paid out amount}
       }}
\usage{
data(PaidReceived)
}
\description{
\code{PaidReceived} Transfers, received and paid out data from Pakistan Social and Living Standards Measurement 2015-16.
}
\references{
\enumerate{
\item Pakistan Bureau of Statistics, Micro data (\url{http://www.pbs.gov.pk/}).
 }
}
\seealso{
\code{\link{Agriculture}}
 , \code{\link{AssetLiability}}
 , \code{\link{Education}}
 , \code{\link{Employment}}
 , \code{\link{Expenditure}}
 , \code{\link{ICT}}
 , \code{\link{Income}}
 , \code{\link{Land}}
 , \code{\link{LiveStock}}
 , \code{\link{NonAgriculture}}
 , \code{\link{OtherIncome}}
 , \code{\link{SurveyInfo}}
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})

Muhammad Arfan Dilber (\email{pbsfsd041@gmail.com})
}
\keyword{datasets}
